Add ADDON definition with subcomponent
