+++ addons/xeh/xeh_contact/CfgVehicles.hpp
@@ -1,22 +1,3 @@
-#include "\x\cba\addons\xeh\script_component.hpp"
-#undef COMPONENT
-#define COMPONENT xeh_compat_contact
-
-class CfgPatches {
-    class ADDON {
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = { "cba_xeh", "A3_Data_F_Contact" };
-        skipWhenMissingDependencies = 1;
-        author = "$STR_CBA_Author";
-        VERSION_CONFIG;
-        // this prevents any patched class from requiring XEH
-        addonRootClass = "A3_Characters_F";
-    };
-};
-class XEH_CLASS_BASE;
-
 class CfgVehicles {
     class B_W_Soldier_F;
     class B_W_Story_Protagonist_01_F: B_W_Soldier_F {
