+++ addons/xeh/xeh_contact/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT contact
+#include "\x\cba\addons\xeh\script_component.hpp"
