+++ addons/xeh/xeh_csla/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT csla
+#include "\x\cba\addons\xeh\script_component.hpp"
