+++ addons/xeh/xeh_sog/CfgVehicles.hpp
@@ -1,22 +1,3 @@
-#include "\x\cba\addons\xeh\script_component.hpp"
-#undef COMPONENT
-#define COMPONENT xeh_compat_sog
-
-class CfgPatches {
-    class ADDON {
-        units[] = {};
-        weapons[] = {};
-        requiredVersion = REQUIRED_VERSION;
-        requiredAddons[] = { "cba_xeh", "loadorder_f_vietnam" };
-        skipWhenMissingDependencies = 1;
-        author = "$STR_CBA_Author";
-        VERSION_CONFIG;
-        // this prevents any patched class from requiring XEH
-        addonRootClass = "A3_Characters_F";
-    };
-};
-class XEH_CLASS_BASE;
-
 class CfgVehicles {
     class vn_object_b_base_02;
     class Land_vn_candle_01: vn_object_b_base_02 {
