+++ addons/xeh/xeh_sog/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT sog
+#include "\x\cba\addons\xeh\script_component.hpp"
