+++ addons/xeh/xeh_ws/script_component.hpp
@@ -1,0 +1,2 @@
+#define SUBCOMPONENT ws
+#include "\x\cba\addons\xeh\script_component.hpp"
